api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
httpx
python-dotenv
pyjwt
greenlet
arq
redis
websockets
neo4j
groq>=0.9.0
google-generativeai>=0.8.0
